Move the types to a dedicated file
