+++ app/components/confirmation-prompt/index.ts
@@ -1,1 +1,2 @@
 export { default } from './component';
+export type { ConfirmationPromptProps } from './types';
