+++ app/components/modal/index.ts
@@ -1,2 +1,2 @@
 export { default } from './component';
-export type { ModalProps } from './component';
+export type { ModalProps } from './types';
